Remove cut from greedy//2
It's quite possible for a caller to want to backtrack into greedy,
thereby giving back some of what it originally consumed.  For example in

    greedy(as),
    greedy(a),
    greedy(b).

The first greedy call has to give back a single "a" so the final two
rules can succeed.

In other words, greedy//2 tries to be as greedy as possible while still
allowing the grammar in which it resides to succeed.