Sketch of a lexer
I want syntax highlighters and code reformatting tools to be able to
work by just tokenizing the code.  I also want the language to have the
simplest syntax that I can manage.  That implies that there should be a
very small set of lexical tokens.

Implementing a small lexer is the best way to see whether I'm staying on
track with those goals.