Simplify tag and key parsing
I realized that I wanted tags and keys to be dash-separated lists of
words.  I already have DCGs for handling lists with separators, so I can
just reuse that code.

They way that I reuse split_tag//2 and split_key//1 suggests that I'm
doing something silly and there's any easier way to do this.  I probably
can't see it because it's late in the day.