Clarify where backtracking won't help
When parsing the goals of a clause or the arguments of a uniline term,
there's no point in backtracking after the first match.  The content we
give back can never match a subsequent rule.

Without this clarification, the parser might expend a lot of useless
work just to discover that it can still make no forward progress.